% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_sdfs.R
\name{query_sdfs}
\alias{query_sdfs}
\alias{pubchem_get_sdf}
\title{Query SDF files of compounds via CID}
\usage{
query_sdfs(cid, dir = "SDFs", curl_cl = NULL, group_number = 50, ...)

pubchem_get_sdf(cid, dir, ...)
}
\description{
Bulk search for compound sdfs via pubchem API.
\url{https://pubchem.ncbi.nlm.nih.gov/rest/pug/compound/cid/2244/SDF}

\code{query_sdfs}: ...

\code{pubchem_get_sdf}: ...
}
\seealso{
Other queries: 
\code{\link{pick_annotation}},
\code{\link{query_classification}()},
\code{\link{query_inchikey}()},
\code{\link{query_iupac}()},
\code{\link{query_synonyms}()}
}
\concept{queries}
