% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Z_class-resp.R
\docType{class}
\name{layerSet-class}
\alias{layerSet-class}
\alias{VIRTUAL_layerSet}
\alias{layerSet}
\alias{layers,layerSet-method}
\alias{layers}
\alias{layers<-,layerSet-method}
\alias{layers<-}
\alias{show,layerSet-method}
\alias{show}
\alias{add_layers,layerSet-method}
\alias{add_layers}
\alias{delete_layers,layerSet,numeric-method}
\alias{delete_layers}
\alias{move_layers,layerSet,numeric,numeric-method}
\alias{move_layers}
\alias{insert_layers,layerSet,numeric-method}
\alias{insert_layers}
\title{Share slots and methods for classes inherite from VIRTUAL_layerSet}
\usage{
\S4method{layers}{layerSet}(x)

\S4method{layers}{layerSet}(x) <- value

\S4method{show}{layerSet}(object)

\S4method{add_layers}{layerSet}(x, ...)

\S4method{delete_layers}{layerSet,numeric}(x, layers)

\S4method{move_layers}{layerSet,numeric,numeric}(x, from, to)

\S4method{insert_layers}{layerSet,numeric}(x, to, ...)
}
\arguments{
\item{x}{object contains slot \code{layers}.}

\item{value}{The value for the slot.}

\item{...}{extra "layer".}

\item{layers}{numeric. The specified "layer" in slot \code{layers}.}

\item{from}{sequence (sequence in list) of "layer" move from.}

\item{to}{sequence (sequence in list) of "layer" move to.}
}
\description{
This VIRTUAL class provides: slot \code{layers} for storing
hierarchical data; and methods for modify slot \code{layers}.

\code{layers}, \code{layers<-}: getter and setter
for the \code{layers} slot of the object.

\code{add_layers}: add extra "layer" into slot \code{layers}.

\code{delete_layers}: delete "layer" in slot \code{layers}.

\code{move_layers}: change the order of "layer" in slot \code{layers}.

\code{insert_layers}: Insert "layers" into the specified
position (sequence) of slot \code{layers}.
}
\section{Slots}{

\describe{
\item{\code{layers}}{list with names.}
}}

\concept{layerSets}
